requirements.txt
fastapi
uvicorn
linkedin-api
python-dotenv
pydantic
slowapi 